[follow_the_gap_v0] Remove (comment out) the LiDAR subscription as it is not needed when we use the obstacles subscription
Logically, it seems that subscription_obstacle_data and subscription_lidar_data should be mutually exclusive. Either we subscribe to LiDAR data (/scan) and use the built-in CreateObstacles function to convert them to obstacles. Or we get obstacles directly from /obstacles subscription (and leave the conversion to the whatever obstacles converter node we are using).