[drive-api] Start the node with publishing calm states
On the new VESC it is a problem when the node publishes 0 values
to Teensy. When the value changes, the PWM is recomputed, and
in the beginning it leads to lower values of PWM than requested.
This results in, e.g., driving in reverse.

Maybe it is a problem of rewritten Teensy FW.